requirements.txt
streamlit
langchain
langchain-core
langchain-openai
langchain-community
faiss-cpu
openai
tiktoken
python-dotenv